Snipe2: electric boogaloo
Other than the more serviceable code, we also have:

* Save cursor position in snipe buffer (got for free by not creating
superfluous buffers)

* Support preselect current open buffer

* Allow opening buffers in splits

* Removed `max_width`, don't think anybody used it (let me know). Now we
have `max_height` which is useful if you want to see only a subset of
buffers

* Buffer list is generated through `:ls` command now, so the results
should be more familiar by default. Path stripping functionality was
taken out, this can be added back as an alternate format function for
menu items.

Next feature I want to add is usign signcolumn to signify meta data
about the buffers, it is more in the spirit of snipes simplicity.